fix: address passthrough review comments
